src/service/chatbot/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
openai==1.3.7
httpx==0.24.1
httpcore==0.17.3
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.7